+++ extensions/guacamole-history-recording-storage/src/main/java/org/apache/guacamole/history/HistoryAuthenticationProvider.java
@@ -19,12 +19,16 @@
 
 package org.apache.guacamole.history;
 
+import java.io.File;
 import org.apache.guacamole.history.user.HistoryUserContext;
 import org.apache.guacamole.GuacamoleException;
+import org.apache.guacamole.environment.Environment;
+import org.apache.guacamole.environment.LocalEnvironment;
 import org.apache.guacamole.net.auth.AbstractAuthenticationProvider;
 import org.apache.guacamole.net.auth.AuthenticatedUser;
 import org.apache.guacamole.net.auth.Credentials;
 import org.apache.guacamole.net.auth.UserContext;
+import org.apache.guacamole.properties.FileGuacamoleProperty;
 
 /**
  * AuthenticationProvider implementation which automatically associates history
@@ -34,6 +38,42 @@ import org.apache.guacamole.net.auth.UserContext;
  */
 public class HistoryAuthenticationProvider extends AbstractAuthenticationProvider {
 
+    /**
+     * The default directory to search for associated session recordings, if
+     * not overridden with the "recording-search-path" property.
+     */
+    private static final File DEFAULT_RECORDING_SEARCH_PATH = new File("/var/lib/guacamole/recordings");
+
+    /**
+     * The directory to search for associated session recordings. By default,
+     * "/var/lib/guacamole/recordings" will be used.
+     */
+    private static final FileGuacamoleProperty RECORDING_SEARCH_PATH = new FileGuacamoleProperty() {
+
+        @Override
+        public String getName() {
+            return "recording-search-path";
+        }
+
+    };
+
+    /**
+     * Returns the directory that should be searched for session recordings
+     * associated with history entries.
+     *
+     * @return
+     *     The directory that should be searched for session recordings
+     *     associated with history entries.
+     *
+     * @throws GuacamoleException
+     *     If the "recording-search-path" property cannot be parsed.
+     */
+    public static File getRecordingSearchPath() throws GuacamoleException {
+        Environment environment = LocalEnvironment.getInstance();
+        return environment.getProperty(RECORDING_SEARCH_PATH,
+                DEFAULT_RECORDING_SEARCH_PATH);
+    }
+    
     @Override
     public String getIdentifier() {
         return "recording-storage";
